refactor: move vote REST API endpoint
